pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aria"
version = "0.1.0"
description = "Automated Regulatory Impact Agent — GraphRAG + multi-agent orchestration for compliance"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "neo4j>=5.27",
    "chromadb>=0.6",
    "litellm>=1.60",
    "pdfplumber>=0.11",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "httpx>=0.28",
    "prometheus-client>=0.21",
    "python-dotenv>=1.0",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]
langgraph = [
    "langgraph>=0.2",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires running Neo4j and ChromaDB",
    "eval: retrieval quality evaluation benchmarks",
]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
